+++ agent/azure-monitor-exporter/src/main/java/com/azure/monitor/opentelemetry/exporter/implementation/quickpulse/QuickPulseDataFetcher.java
@@ -183,7 +183,7 @@ class QuickPulseDataFetcher {
           new QuickPulseMetrics(
               "\\ApplicationInsights\\Dependency Call Duration",
               counters.rddsDuration / counters.rdds,
-              counters.rdds));
+              (int) counters.rdds));
     }
     metricsList.add(
         new QuickPulseMetrics(
